Implement baseline alignment in layout engine

+++ content/handlers/html/layout_flex.c
@@ -40,6 +40,7 @@
 #include "html/box.h"
 #include "html/html.h"
 #include "html/private.h"
+#include "netsurf/inttypes.h"
 #include "html/box_inspect.h"
 #include "html/layout_internal.h"
 
@@ -773,7 +774,7 @@ static bool layout_flex__resolve_line(
 
 	NSLOG(flex, DEEPDEBUG, "box %p: line %" PRIsizet ": first: %" PRIsizet ", count: %" PRIsizet "",
 			(void *)ctx->flex, (size_t)(line - ctx->line.data),
-			line->first, line->count);
+			(size_t)line->first, (size_t)line->count);
 	NSLOG(flex, DEEPDEBUG, "Line main_size: %i, available_main: %i",
 			line->main_size, available_main);
 
@@ -1018,8 +1019,8 @@ static bool layout_flex__collect_items_into_lines(
 
 		NSLOG(flex, DEEPDEBUG, "flex-container: %p: "
 				"fitted: %" PRIsizet " (total: %" PRIsizet "/%" PRIsizet ")",
-				(void *)ctx->flex, line->count,
-				pos, ctx->item.count);
+				(void *)ctx->flex, (size_t)line->count,
+				(size_t)pos, (size_t)ctx->item.count);
 
 		if (!layout_flex__resolve_line(ctx, line)) {
 			return false;
